fix: prevent harmless error at build time

+++ copy-libpanel.sh
@@ -5,5 +5,7 @@ LIBPANEL_DST=./dist/libs/libpanel
 
 mkdir -p $(dirname $LIBPANEL_DST)
 
-rm -r $LIBPANEL_DST
+if [ -d $LIBPANEL_DST ]; then
+	rm -r $LIBPANEL_DST
+fi
 cp -r $LIBPANEL_SRC $LIBPANEL_DST
